Refactor the eggnog module
