+++ q2_moshpit/eggnog/__init__.py
@@ -5,20 +5,16 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
-from ._method import (
-    eggnog_diamond_search, _eggnog_diamond_search,
-    eggnog_hmmer_search, _eggnog_hmmer_search,
-    eggnog_annotate, _eggnog_annotate, _eggnog_feature_table,
-    _symlink_files_to_target_dir, _search_runner, _eggnog_search
+from .orthologs import (
+    eggnog_diamond_search, _eggnog_diamond_search, eggnog_hmmer_search,
+    _eggnog_hmmer_search, _eggnog_feature_table, _search_runner,
+    _eggnog_search
 )
-from ._dbs import (
+from .annotation import eggnog_annotate, _eggnog_annotate
+from .dbs import (
     fetch_eggnog_db, fetch_diamond_db, build_custom_diamond_db,
     fetch_eggnog_proteins, build_eggnog_diamond_db, fetch_ncbi_taxonomy,
     fetch_eggnog_hmmer_db
-)
-from q2_moshpit.eggnog._type import EggnogHmmerIdmap
-from q2_moshpit.eggnog._format import (
-    EggnogHmmerIdmapDirectoryFmt, EggnogHmmerIdmapFileFmt
 )
 
 
@@ -27,8 +23,6 @@ __all__ = [
     '_eggnog_feature_table', 'fetch_eggnog_db', 'fetch_diamond_db',
     'build_custom_diamond_db', 'fetch_eggnog_proteins',
     'build_eggnog_diamond_db', 'fetch_ncbi_taxonomy', '_eggnog_annotate',
-    'fetch_eggnog_hmmer_db', 'EggnogHmmerIdmapFileFmt', 'EggnogHmmerIdmap',
-    'EggnogHmmerIdmapDirectoryFmt', 'eggnog_hmmer_search',
-    '_eggnog_hmmer_search', '_symlink_files_to_target_dir', '_search_runner',
-    '_eggnog_search'
+    'fetch_eggnog_hmmer_db', 'eggnog_hmmer_search', '_eggnog_hmmer_search',
+    '_search_runner', '_eggnog_search'
 ]
