+++ q2_moshpit/eggnog/orthologs/__init__.py
@@ -1,0 +1,24 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2022-2023, QIIME 2 development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+from .common import (
+    _run_eggnog_search_pipeline, _eggnog_search, _search_runner,
+    _eggnog_feature_table
+)
+from .diamond import (
+    _eggnog_diamond_search, eggnog_diamond_search
+)
+from .hmmer import (
+    _eggnog_hmmer_search, eggnog_hmmer_search
+)
+
+
+__all__ = [
+    '_run_eggnog_search_pipeline', '_eggnog_search', '_search_runner',
+    '_eggnog_diamond_search', 'eggnog_diamond_search',
+    '_eggnog_hmmer_search', 'eggnog_hmmer_search', '_eggnog_feature_table'
+]
