+++ q2_moshpit/eggnog/types/__init__.py
@@ -1,0 +1,15 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2022-2023, QIIME 2 development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+from ._type import EggnogHmmerIdmap
+from ._format import EggnogHmmerIdmapFileFmt, EggnogHmmerIdmapDirectoryFmt
+
+
+__all__ = [
+    'EggnogHmmerIdmapFileFmt', 'EggnogHmmerIdmapDirectoryFmt',
+    'EggnogHmmerIdmap'
+]
